+++ app/models/concerns/federails/data_entity.rb
@@ -182,7 +182,7 @@ module Federails
 
       before_validation :set_federails_actor
       after_create -> { create_federails_activity 'Create' }
-      after_update -> { create_federails_activity 'Update' }, :federails_tombstoned?
+      after_update -> { create_federails_activity 'Update' }, unless: :federails_tombstoned?
       after_destroy -> { create_federails_activity 'Delete' }
     end