+++ CHANGELOG.md
@@ -54,6 +54,7 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 - Client:
   - Don't fail in `follower` partial when user is not signed-in
   - Require `uri` parameter in new following action 
+  - Strip parameter value when searching for actors
 
 ## [0.6.2] 2025-04-18