+++ CHANGELOG.md
@@ -55,6 +55,8 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
   - Don't fail in `follower` partial when user is not signed-in
   - Require `uri` parameter in new following action 
   - Strip parameter value when searching for actors
+- Server:
+  - Fix nodeinfo scheme in schema URL
 
 ## [0.6.2] 2025-04-18